requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
PyJWT==0.2.1
PyMySQL==0.6.2
PyYAML==3.11
Werkzeug==0.9.6
amqp==1.4.5
anyjson==0.3.3
billiard==3.3.0.18
celery==3.1.13
httplib2==0.9
itsdangerous==0.24
kombu==3.0.21
mwoauth==0.2.2
oauthlib==0.6.3
pytz==2014.4
redis==2.10.1
requests==2.3.0
requests-oauthlib==0.4.1
six==1.7.3
wsgiref==0.1.2
SQLAlchemy==0.9.7
